+++ odin-core/src/main/java/com/purplepip/odin/sequence/SameTimeUnitTickConverter.java
@@ -44,7 +44,7 @@ public class SameTimeUnitTickConverter extends AbstractTickConverter {
   }
 
   @Override
-  protected double getTimeWithBeatBasedTimeUnits(Direction direction, long time) {
+  protected double getTimeWithBeatBasedTimeUnits(Direction direction, double time) {
     if (getSourceTick().getTimeUnit() == TimeUnit.BEAT) {
       return direction.scaleTime(time);
     }
@@ -52,7 +52,7 @@ public class SameTimeUnitTickConverter extends AbstractTickConverter {
   }
 
   @Override
-  protected double getTimeWithMicrosecondBasedTimeUnits(Direction direction, long time) {
+  protected double getTimeWithMicrosecondBasedTimeUnits(Direction direction, double time) {
     if (getSourceTick().getTimeUnit() == TimeUnit.MICROSECOND) {
       return direction.scaleTime(time);
     }
